feat: Add ApiEnum and apiEnumValueOfOrNull helper methods
Now that we write enums with PascalCase we can't always use the `enumValueOfOrNull` and base the recognition on the name of the enum value. So I added a helper method to easily get an enum instance for any enum that extends the ApiEnum interface. The interface only requires the children to implement an `apiValue` field that will be used to match the input string. The helper method is pretty similar to the existing `enumValueOfOrNull` but with a different name.

Downgrading androidxCore was necessary to use the module on Mail for the time being.